Adds identified license report format
