+++ src/api/report-service.ts
@@ -25,6 +25,4 @@ class ReportService extends BaseService {
   }
 }
 
-
-
 export const reportService = new ReportService();
