+++ src/renderer/features/workbench/components/AppBar/AppBar.tsx
@@ -219,18 +219,15 @@ const Export = ({ state }) => {
       >
         <MenuItem disabled={state.progress === 0} onClick={() => onExport(FormatVersion.CSV)}>
           CSV
-        </MenuItem>
-        {/* <MenuItem disabled={state.progress === 0} onClick={() => onExport(FormatVersion.SPDX20)}>
-          SPDX 2.0
-        </MenuItem> */}
-        {/* <MenuItem disabled={state.progress === 0} onClick={() => onExport(FormatVersion.SPDXLITE)}>
-          SPDX Lite
-        </MenuItem> */}
+        </MenuItem> 
         <MenuItem disabled={state.progress === 0} onClick={() => onExport(FormatVersion.SPDXLITEJSON)}>
           SPDX Lite
         </MenuItem>
         <MenuItem onClick={() => onExport(FormatVersion.WFP)}>WFP</MenuItem>
         <MenuItem onClick={() => onExport(FormatVersion.RAW)}>RAW</MenuItem>
+        <MenuItem disabled={state.progress === 0} onClick={() => onExport(FormatVersion.LICENSEREPORT)}>
+          Identified License Report
+        </MenuItem>
       </Menu>
     </div>
   );
